Animate light

+++ src/main/java/yeyu/dynamiclights/client/LightFunction.java
@@ -2,11 +2,11 @@ package yeyu.dynamiclights.client;
 
 import java.util.function.BiFunction;
 
-public enum LightFunction implements BiFunction<Float, Integer, Float> {
+public enum LightFunction implements BiFunction<Float, Float, Float> {
     QUADRATIC;
 
     @Override
-    public Float apply(Float squaredDistance, Integer maxLight) {
+    public Float apply(Float squaredDistance, Float maxLight) {
         return Math.max(0, Math.min((-(DynamicLightsOption.getCurrentLightMultiplier() * (squaredDistance))) + 18, 14) - (14 - maxLight));
     }
 }
